Silence warnings on JRuby
While I cannot be sure that the bindings collection will work properly
outside of the non-JITted JRuby modes (because of the optimizations
JRuby does, local variables may end up on the stack and changing them in
a binding may lead to unexpected results), lets document that and not
warn every time. The test suite runs fine even without --debug or --dev
or interpreted mode. This can change, though.