CRuby: Don't monkey patch Exception#set_backtrace
I can do that with the tracepoint API in Ruby 2.0 and above and
rb_add_event_hook in Ruby 1.9.3.

This is a great thing in general, as I can drop the core header files
for 2.0 and above. I can't afford that luxury for 1.9.3, though.

As a side effect I have collected all of the code in one place. Its
kinda hard to manage with all the 1.9.3 hacks otherwise.